+++ package.json
@@ -16,6 +16,7 @@
     "@emotion/server": "^11.10.0",
     "@emotion/styled": "^11.10.0",
     "@jest/types": "^28.1.3",
+    "@mui/icons-material": "^5.10.2",
     "@mui/material": "^5.10.0",
     "@reduxjs/toolkit": "^1.8.4",
     "@testing-library/react": "^13.3.0",
