--- pages/visitors/[visitorId].js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { useRouter } from 'next/router';
-
-const VisitorId = () => {
-  const router = useRouter();
-  const visitorId = router.query.visitorId;
-
-  return <h2>Detail about the visitor #{visitorId}</h2>;
-};
-
-export default VisitorId;
