--- pages/visitors/index.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const Visitor = () => {
-  return (
-    <>
-      <h2>visitor 1</h2>
-      <h2>visitor 2</h2>
-      <h2>visitor 3</h2>
-    </>
-  );
-};
-
-export default Visitor;
